SubscriptionOptions? is now nullable on StoreProduct and find defaul offer moved into SubscriptionOptions
