+++ common/src/test/java/com/revenuecat/purchases/common/BackendTest.kt
@@ -280,7 +280,7 @@ class BackendTest {
 
     @Test
     fun `postReceipt passes pricing phases as maps in body`() {
-        val subscriptionOption = storeProduct.subscriptionOptions[0]
+        val subscriptionOption = storeProduct.subscriptionOptions!!.first()
         val receiptInfo = ReceiptInfo(
             productIDs = productIDs,
             storeProduct = storeProduct,
@@ -591,7 +591,7 @@ class BackendTest {
             storeProduct = storeProduct
         )
 
-        val originalSubscriptionOption = storeProduct.subscriptionOptions[0]
+        val originalSubscriptionOption = storeProduct.subscriptionOptions!!.first()
         val originalDuration = originalSubscriptionOption.pricingPhases[0].billingPeriod.iso8601
         val subscriptionOption = stubSubscriptionOption(originalSubscriptionOption.id, originalDuration + "a")
         val storeProduct2 = stubStoreProduct(
