+++ public/src/main/java/com/revenuecat/purchases/models/GoogleStoreProduct.kt
@@ -16,8 +16,8 @@ data class GoogleStoreProduct(
     override val title: String,
     override val description: String,
     override val period: Period?,
-    override val subscriptionOptions: SubscriptionOptions,
-    override val defaultOption: GoogleSubscriptionOption?,
+    override val subscriptionOptions: SubscriptionOptions?,
+    override val defaultOption: SubscriptionOption?,
     val productDetails: @RawValue ProductDetails // TODO parcelize?
 ) : StoreProduct, Parcelable {
