+++ public/src/main/java/com/revenuecat/purchases/models/SubscriptionOptions.kt
@@ -1,12 +1,19 @@
 package com.revenuecat.purchases.models
 
 import android.os.Parcelable
+import androidx.annotation.VisibleForTesting
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
 class SubscriptionOptions(
     private val subscriptionOptions: List<SubscriptionOption>
     ) : List<SubscriptionOption> by subscriptionOptions, Parcelable {
+
+    /**
+     * Finds the base plan [SubscriptionOption].
+     */
+    val basePlan: SubscriptionOption?
+        get() = this.firstOrNull { it.isBasePlan }
 
     /**
      * Finds the first [SubscriptionOption] with a free trial [PricingPhase].
@@ -28,18 +35,69 @@ class SubscriptionOptions(
         return this.filter { it.tags.contains(tag) }
     }
 
+    fun findDefaultOffer(): SubscriptionOption? {
+        val basePlan = this.firstOrNull { it.isBasePlan } ?: return null
+
+        val validOffers = this
+            .filter { !it.isBasePlan }
+            .filter { !it.tags.contains("rc-ignore-default-offer") }
+
+        return findLongestFreeTrial(validOffers) ?: findLowestNonFreeOffer(validOffers) ?: basePlan
+    }
+
+    private fun findLongestFreeTrial(offers: List<SubscriptionOption>): SubscriptionOption? {
+        return offers.mapNotNull { offer ->
+            offer.freePricingPhase?.let { pricingPhase ->
+                Pair(offer, billingPeriodToDays(pricingPhase.billingPeriod))
+            }
+        }.maxByOrNull { it.second }?.first
+    }
+
+    private fun findLowestNonFreeOffer(offers: List<SubscriptionOption>): SubscriptionOption? {
+        return offers.mapNotNull { offer ->
+            offer.nonFreePricingPhase?.let { pricingPhase ->
+                Pair(offer, pricingPhase.price.amountMicros)
+            }
+        }.minByOrNull { it.second }?.first
+    }
+
+    private val SubscriptionOption.freePricingPhase: PricingPhase?
+        get() = pricingPhases.firstOrNull()?.takeIf { it.price.amountMicros == 0L }
+
+    private val SubscriptionOption.nonFreePricingPhase: PricingPhase?
+        get() = pricingPhases.firstOrNull()?.takeIf { it.price.amountMicros > 0L }
+
+    @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+    internal fun billingPeriodToDays(period: Period): Int {
+        val days = DAYS_IN_UNIT[period.unit] ?: 0
+        return period.value * days
+    }
+
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (javaClass != other?.javaClass) return false
 
-        other as SubscriptionOptions
+        (other as? SubscriptionOptions)?.let {
+            if (listOf(this.subscriptionOptions) != listOf(other.subscriptionOptions)) return false
 
-        if (listOf(this.subscriptionOptions) != listOf(other.subscriptionOptions)) return false
-
-        return true
+            return true
+        } ?: run {
+            return false
+        }
     }
 
     override fun hashCode(): Int {
         return listOf(this.subscriptionOptions).hashCode()
     }
 }
+
+private const val DAYS_IN_DAY = 1
+private const val DAYS_IN_WEEK = 7
+private const val DAYS_IN_MONTH = 30
+private const val DAYS_IN_YEAR = 365
+private val DAYS_IN_UNIT = mapOf(
+    Period.Unit.DAY to DAYS_IN_DAY,
+    Period.Unit.WEEK to DAYS_IN_WEEK,
+    Period.Unit.MONTH to DAYS_IN_MONTH,
+    Period.Unit.YEAR to DAYS_IN_YEAR,
+)
